refactor: 妖精がfairySpeechAPIを介して喋るように
