+++ src/main/scala/com/github/unchama/seichiassist/subsystems/vote/subsystems/fairy/System.scala
@@ -3,10 +3,7 @@ package com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy
 import cats.data.Kleisli
 import cats.effect.{ConcurrentEffect, IO, SyncIO}
 import com.github.unchama.concurrent.RepeatingTaskContext
-import com.github.unchama.datarepository.bukkit.player.{
-  BukkitRepositoryControls,
-  PlayerDataRepository
-}
+import com.github.unchama.datarepository.bukkit.player.BukkitRepositoryControls
 import com.github.unchama.datarepository.template.RepositoryDefinition
 import com.github.unchama.minecraft.actions.OnMinecraftServerThread
 import com.github.unchama.seichiassist.concurrent.PluginExecutionContexts.timer
@@ -23,17 +20,20 @@ import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.applicat
 import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.bukkit.actions.BukkitSummonFairy
 import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.bukkit.listeners.FairyPlayerJoinGreeter
 import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.bukkit.routines.BukkitFairyRoutine
-import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.bukkit.BukkitFairySummonRequest
+import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.bukkit.{
+  BukkitFairySpeech,
+  BukkitFairySummonRequest
+}
 import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.domain.property._
+import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.domain.speech.FairySpeech
 import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.domain.{
   FairyPersistence,
   FairySpawnRequestErrorOrSpawn,
   FairySummonRequest
 }
 import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.infrastructure.JdbcFairyPersistence
-import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairyspeech.bukkit.{BukkitFairySpeech, BukkitFairySpeechGateway}
-import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairyspeech.domain.{FairySpeech, FairySpeechGateway}
-import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairyspeech.service.FairySpeechService
+import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairyspeech.bukkit.BukkitFairySpeechGateway
+import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairyspeech.domain.FairySpeechGateway
 import org.bukkit.entity.Player
 import org.bukkit.event.Listener
 import org.bukkit.inventory.ItemStack
@@ -51,6 +51,10 @@ object System {
     voteAPI: VoteAPI[IO, Player],
     manaApi: ManaApi[IO, SyncIO, Player],
     mineStackAPI: MineStackAPI[IO, Player, ItemStack],
+    fairySpeechAPI: com.github.unchama.seichiassist.subsystems.vote.subsystems.fairyspeech.FairySpeechAPI[
+      IO,
+      Player
+    ],
     repeatingTaskContext: RepeatingTaskContext,
     concurrentEffect: ConcurrentEffect[IO],
     minecraftServerThread: OnMinecraftServerThread[IO]
@@ -58,9 +62,7 @@ object System {
     implicit val persistence: FairyPersistence[IO] = new JdbcFairyPersistence[IO]
     implicit val fairySpeechGatewayProvider: Player => FairySpeechGateway[SyncIO] =
       new BukkitFairySpeechGateway[SyncIO](_)
-    val fairySpeechProvider
-      : PlayerDataRepository[FairySpeechService[SyncIO]] => FairySpeech[IO, Player] =
-      fairySpeechService => new BukkitFairySpeech[IO, SyncIO](fairySpeechService, persistence)
+    implicit val fairySpeechProvider: FairySpeech[IO, Player] = new BukkitFairySpeech[IO]
 
     for {
       speechServiceRepositoryControls <- BukkitRepositoryControls.createHandles(
@@ -71,9 +73,7 @@ object System {
             SpeechServiceRepositoryDefinitions.finalization[SyncIO, Player]
           )
       )
-      _fairyRoutine = new BukkitFairyRoutine(
-        fairySpeechProvider(speechServiceRepositoryControls.repository)
-      )
+      _fairyRoutine = new BukkitFairyRoutine(fairySpeechProvider)
       fairyRecoveryRoutineFiberRepositoryControls <- BukkitRepositoryControls.createHandles(
         RepositoryDefinition
           .Phased
@@ -85,11 +85,6 @@ object System {
       )
     } yield {
       new System[IO, SyncIO, Player] {
-        val fairySpeechServiceRepository: PlayerDataRepository[FairySpeechService[SyncIO]] =
-          speechServiceRepositoryControls.repository
-        implicit val fairySpeech: FairySpeech[IO, Player] = fairySpeechProvider(
-          fairySpeechServiceRepository
-        )
         implicit val summonFairy: SummonFairy[IO, Player] =
           new BukkitSummonFairy[IO, SyncIO]
         val summonRequest: FairySummonRequest[IO, Player] =
@@ -132,7 +127,7 @@ object System {
               persistence.totalConsumedAppleAmount
 
             override def sendDisappearTimeToChat: Kleisli[IO, Player, Unit] = Kleisli {
-              player => fairySpeech.speechEndTime(player)
+              player => fairySpeechProvider.speechEndTime(player)
             }
 
             override def fairySummonRequest
